rename tests dir
