+++ integration-tests/jenkins-image/Dockerfile
