+++ integration-tests/jenkins-image/init-cloud.groovy
