+++ integration-tests/jenkins-image/init-long-job.groovy
