+++ integration-tests/jenkins-image/init-short-job.groovy
