+++ integration-tests/jenkins-image/plugins.txt
