+++ integration-tests/jenkins-image/stop-jenkins.sh
